:bug: avoid short writes when writing out values
The previous code was not using write_all when writing out values to
the given stream. As a result, it was possible to incompletely write
out the value. Such a sort write would typically result in the client
blocking indefinitely as it tried to read a server response while the
server was still waiting for data.

We encountered this when moving to the TLS transport since the
underlying openssl library uses a 16KB buffer by default and some of
our value bodies were over 16KB.

Signed-off-by: Steven Danna <steve@chef.io>