:building_construction: Add examples directory
To get this to work, we also need to move the code from src/ to its own
package in memcache. This is because we're adding a complex example
with axum, which includes its own Cargo.toml. This has other effects: it
means that you can't add single file examples anymore. It must be a
crate. This is the same decision used in the Axum repo, and is just a
current limitation of cargo.